INFO2df {vcfR}	R Documentation

Reformat INFO data as a data.frame

Description

Reformat INFO data as a data.frame and handle class when possible.

Usage

INFO2df(x)

metaINFO2df(x, field = "INFO")

Arguments

x
an object of class vcfR or chromR.

field
should either the INFo or FORMAT data be returned?

Details

The INFO column of VCF data contains descriptors for each variant. Because this column may contain many comma delimited descriptors it may be difficult to interpret. The function INFO2df converts the data into a data.frame. The function metaINFO2df extracts the information in the meta section that describes the INFO descriptors. This function is called by INFO2df to help it handle the class of the data.

Value

A data.frame

Examples

data(vcfR_test)
metaINFO2df(vcfR_test)
getINFO(vcfR_test)
INFO2df(vcfR_test)

[Package vcfR version 1.15.0 Index]