check_keys {vcfR}	R Documentation

Check that INFO and FORMAT keys are unique

Description

The INFO and FORMAT columns contain information in key-value pairs. If for some reason a key is not unique it will create issues in retrieving this information. This function checks the keys defined in the meta section to make sure they are unique. Note that it does not actually check the INFO and FORMAT columns, just their definitions in the meta section. This is because each variant can have different information in their INFO and META cells. Checking these on large files will therefore come with a performance cost.

Usage

check_keys(x)

Arguments

x
an oblect of class vcfR

See Also

queryMETA()

Examples

data(vcfR_test)
check_keys(vcfR_test)
queryMETA(vcfR_test)
queryMETA(vcfR_test, element = 'DP')
# Note that DP occurs as unique in INFO and FORMAT but they may be different.

[Package vcfR version 1.15.0 Index]