chromR-class {vcfR}	R Documentation

chromR class

Description

A class for representing chromosomes (or supercontigs, contigs, scaffolds, etc.).

Details

Defines a class for chromosomal or contig data. This

This object has a number of slots.

name name of the object (character)

len length of the sequence (integer)

window_size window size for windowing analyses (integer)

seq object of class ape::DNAbin

vcf object of class vcfR

ann annotation data in a gff-like data.frame

var.info a data.frame containing information on variants

win.info a data.frame containing information on windows

seq.info a list containing information on the sequence

The seq slot contains an object of class ape::DNAbin. A DNAbin object is typically either a matrix or list of DNAbin objects. The matrix form appears to be better behaved than the list form. Because of this behavior this slot should be the matrix form. When this slot is not populated it is of class "NULL" instead of "DNAbin". Note that characters need to be lower case when inserted into an object of class DNAbin. The function tolower can facilitate this.

The vcf slot is an object of class vcfR vcfR-class.

The ann slot is a data.frame containing gff format data. When this slot is not populated it has nrows equal to zero.

The var.info slot contains a data.frame containing information about variants. Every row of this data.frame is a variant. Columns will typically contain the chromosome name, the position of the variant (POS), the mask as well as any other per variant information.

The win.info slot contains a data.frame containing information about windows. For example, window, start, end, length, A, C, G, T, N, other, variants and genic fields are stored here.

The seq.info slot is a list containing two matrices. The first matrix describes rectangles for called nucleotides and the second describes rectangles for 'N' calls. Within each matrix, the first column indicates the start position and the second column indicates the end position of each rectangle.

See Also

vcfR-class, DNAbin, VCF specification gff3 format

[Package vcfR version 1.15.0 Index]