chromR functions {vcfR}	R Documentation

chromR_functions

Description

Functions which act on chromR objects

Usage

masker(
  x,
  min_QUAL = 1,
  min_DP = 1,
  max_DP = 10000,
  min_MQ = 20,
  max_MQ = 100,
  preserve = FALSE,
  ...
)

variant.table(x)

win.table(x)

Arguments

x
object of class chromR

min_QUAL
minimum variant quality

min_DP
minimum cumulative depth

max_DP
maximum cumulative depth

min_MQ
minimum mapping quality

max_MQ
maximum mapping quality

preserve
a logical indicating whether or not to preserve the state of the current mask field. Defaults to FALSE

...
arguments to be passed to methods

Details

The function masker creates a logical vector that determines which variants are masked. By masking certain variants, instead of deleting them, it preserves the dimensions of the data structure until a change needs to be committed. Variants can be masked based on the value of the QUAL column of the vcf object. Experience seems to show that this value is either at its maximum (999) or a rather low value. The maximum and minimum sequence depth can also be used (mindp and maxdp). The default is to mask all variants with depths of less than the 0.25 quantile and greater than the 0.75 quantile (these are also known as the lower and upper quartile). The minimum and maximum mapping qualities (minmq, maxmq) can also be used.

This vector is stored in the var.info$mask slot of a chromR object.

The function variant.table creates a data.frame containing information about variants.

The funciton win.table

[Package vcfR version 1.15.0 Index]