chromR2vcfR {vcfR}	R Documentation

Convert chrom objects to vcfR objects

Description

Convert chrom objects to vcfR objects.

Usage

chromR2vcfR(x, use.mask = FALSE)

Arguments

x
Object of class chrom

use.mask
Logical, determine if mask from chrom object should be used to subset vcf data

Details

The chrom object is subset and recast as a vcfR object. When use.mask is set to TRUE (the default), the object is subset to only the variants (rows) indicated to include by the mask. When use.mask is set to FALSE, all variants (rows) from the chrom object are included in the new vcfR object.

Value

Returns an object of class vcfR.

[Package vcfR version 1.15.0 Index]