chromo_plot {vcfR}	R Documentation

Plot chromR object

Description

plot chromR objects

Usage

chromo(
  chrom,
  boxp = TRUE,
  dp.alpha = TRUE,
  chrom.s = 1,
  chrom.e = NULL,
  drlist1 = NULL,
  drlist2 = NULL,
  drlist3 = NULL,
  ...
)

chromoqc(chrom, boxp = TRUE, dp.alpha = 255, ...)

Arguments

chrom
an object of class chrom.

boxp
logical specifying whether marginal boxplots should be plotted [T/F].

dp.alpha
degree of transparency applied to points in dot plots [0-255].

chrom.s
start position for the chromosome. (Deprecated. use xlim)

chrom.e
end position for the chromosome. (Deprecated. use xlim)

drlist1
a named list containing elements to create a drplot

drlist2
a named list containing elements to create a drplot

drlist3
a named list containing elements to create a drplot

...
arguments to be passed to other methods.

Details

Each drlist parameter is a list containing elements necessarry to plot a dr.plot. This list should contain up to seven elements named title, dmat, rlist, dcol, rcol, rbcol and bwcol. These elements are documented in the dr.plot page where they are presented as individual parameters. The one exception is bwcol which is a vector of colors for the marginal box and whisker plot. This is provided so that different colors may be used in the dot plot and the box and whisker plot. For example, transparency may be desired in the dot plot but not the box and whisker plot. When one (or more) of these elements is omitted an attempt to use default values is made.

Value

Returns an invisible NULL.

See Also

dr.plot

[Package vcfR version 1.15.0 Index]