dr.plot elements {vcfR}	R Documentation

dr.plot elements

Description

Plot chromR objects and their components

Usage

dr.plot(
  dmat = NULL,
  rlst = NULL,
  chrom.s = 1,
  chrom.e = NULL,
  title = NULL,
  hline = NULL,
  dcol = NULL,
  rcol = NULL,
  rbcol = NULL,
  ...
)

null.plot()

Arguments

dmat
a numeric matrix for dot plots where the first column is position (POS) and subsequent columns are y-values.

rlst
a list containing numeric matrices containing rectangle coordinates.

chrom.s
start position for the chromosome

chrom.e
end position for the chromosome

title
optional string to be used for the plot title.

hline
vector of positions to be used for horizontal lines.

dcol
vector of colors to be used for dot plots.

rcol
vector of colors to be used for rectangle plots.

rbcol
vector of colors to be used for rectangle borders.

...
arguments to be passed to other methods.

Details

Plot details The parameter rlist is list of numeric matrices containing rectangle coordinates. The first column of each matrix is the left positions, the second column is the bottom coordinates, the third column is the right coordinates and the fourth column is the top coordinates.

Value

Returns the y-axis minimum and maximum values invisibly.

See Also

rect chromo

[Package vcfR version 1.15.0 Index]