write.fasta {vcfR}	R Documentation

Create fasta format output

Description

Generate fasta format output

Usage

write.fasta(
  x,
  file = "",
  rowlength = 80,
  tolower = TRUE,
  verbose = TRUE,
  APPEND = FALSE,
  depr = TRUE
)

Arguments

x
object of class chromR

file
name for output file

rowlength
number of characters each row should not exceed

tolower
convert all characters to lowercase (T/F)

verbose
should verbose output be generated (T/F)

APPEND
should data be appended to an existing file (T/F)

depr
logical (T/F), this function has been deprecated, set to FALSE to override.

Details

The function write_fasta takes an object of class chromR and writes it to a fasta.gz (gzipped text) format file. The sequence in the seq slot of the chromR object is used to fill in the invariant sites. The parameter 'tolower', when set to TRUE, converts all the characters in teh sequence to lower case. This is important because some software, such as ape::DNAbin, requires sequences to be in lower case.

[Package vcfR version 1.15.0 Index]