getFIX {vcfR}	R Documentation

Get elements from the fixed region of a VCF file

Description

Both chromR objects and vcfR objects contain a region with fixed variables. These accessors allow you to isolate these variables from these objects.

Usage

getFIX(x, getINFO = FALSE)

getCHROM(x)

getPOS(x)

getQUAL(x)

getALT(x)

getREF(x)

getID(x)

getFILTER(x)

getINFO(x)

Arguments

x
a vcfR or chromR object

getINFO
logical specifying whether getFIX should return the INFO column

Value

a vector or data frame

Examples

library("vcfR")
data("vcfR_example")
data("chromR_example")
getFIX(vcf) %>% head
getFIX(chrom) %>% head

getCHROM(vcf) %>% head
getCHROM(chrom) %>% head

getPOS(vcf) %>% head
getPOS(chrom) %>% head

getID(vcf) %>% head
getID(chrom) %>% head

getREF(vcf) %>% head
getREF(chrom) %>% head

getALT(vcf) %>% head
getALT(chrom) %>% head

getQUAL(vcf) %>% head
getQUAL(chrom) %>% head

getFILTER(vcf) %>% head
getFILTER(chrom) %>% head

getINFO(vcf) %>% head
getINFO(chrom) %>% head

[Package vcfR version 1.15.0 Index]