queryMETA {vcfR}	R Documentation

Query the META section of VCF data

Description

Query the META section of VCF data for information about acronyms.

Usage

queryMETA(x, element = NULL, nice = TRUE)

Arguments

x
an object of class vcfR or chromR.

element
an acronym to search for in the META portion of the VCF data.

nice
logical indicating whether to format the data in a 'nice' manner.

Details

The META portion of VCF data defines acronyms that are used elsewhere in the data. In order to better understand these acronyms they should be referenced. This function facilitates looking up of acronyms to present their relevant information. When 'element' is 'NULL' (the default), all acronyms from the META region are returned. When 'element' is specified an attempt is made to return information about the provided element. The function grep is used to perform this query. If 'nice' is set to FALSE then the data is presented as it was in the file. If 'nice' is set to TRUE the data is processed to make it appear more 'nice'.

See Also

grep, regex.

Examples

data(vcfR_test)
queryMETA(vcfR_test)
queryMETA(vcfR_test, element = "DP")

[Package vcfR version 1.15.0 Index]