write.var.info {vcfR}	R Documentation

Write summary tables from chromR objects

Description

Write summary tables from chromR objects.

Usage

write.var.info(x, file = "", mask = FALSE, APPEND = FALSE)

write.win.info(x, file = "", APPEND = FALSE)

Arguments

x
An object of class chromR

file
A filename for the output file

mask
logical vector indicating rows to use

APPEND
logical indicating whether to append to existing file (omitting the header) or write a new file

Details

The function write.var.info takes the variant information table from a chromR object and writes it as a comma delimited file.

The function write.win.info takes the window information table from a chromR object and writes it as a comma delimited file.

See Also

write.vcf

[Package vcfR version 1.15.0 Index]