vcfR-class {vcfR}	R Documentation

vcfR class

Description

An S4 class for storing VCF data.

Details

Defines a class for variant call format data. A vcfR object contains three slots. The first slot is a character vector which holds the meta data. The second slot holds an eight column matrix to hold the fixed data. The third slot is a matrix which holds the genotype data. The genotype data is optional according to the VCF definition. When it is missing the gt slot should consist of a character matrix with zero rows and columns.

See vignette('vcf_data') for more information. See the VCF specification for the file specification.

Slots

meta

character vector for the meta information

fix

matrix for the fixed information

gt

matrix for the genotype information

[Package vcfR version 1.15.0 Index]