vcfR_test {vcfR}	R Documentation

Test data for vcfR.

Description

A test file containing a diversity of examples intended to test functionality.

Format

A vcfR object

Details

vcfR_test vcfR object

This data set began as the example (section 1.1) from The Variant Call Format Specification VCFv4.3 . This data consisted of 3 samples and 5 variants. As I encounter examples that challenge the code in vcfR they can be added to this data set.

Examples

data(vcfR_test)

## Not run: 
  # When I add data it can be saved with this command.
  save(vcfR_test, file="data/vcfR_test.RData") 
  
## End(Not run)

[Package vcfR version 1.15.0 Index]