vep {vcfR}	R Documentation

Example data from the Variant Effect Predictor (VEP).

Description

Example data to use with unit tests.

Format

A vcfR object

Details

vep vcfR object

Output from the VEP may include values with multiple equals signs. This does not appear to conform with the VCF specification (at the time of writing this VCF v4.3). But it appears fairly easy to accomodate. This example data can be used to make unit tests to validate functionality.

Examples

data(vep)
vcfR2tidy(vep, info_only = TRUE)$fix

[Package vcfR version 1.15.0 Index]